requirements.txt
faiss-cpu
sentence-transformers
PyPDF2
openai
numpy
streamlit
python-dotenv
